% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{data_checks}
\alias{data_checks}
\title{data_checks}
\usage{
data_checks(C_max, group_contbn, ped, fams, candidates, parents_max, method)
}
\value{
Error message if error identified
}
\description{
data_checks
}
