% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fams_example.R
\docType{data}
\name{fams_example}
\alias{fams_example}
\title{fams_example data for small worked example in Hamilton (2020) Optimal  contributions in highly fecund species with overlapping generations}
\format{
Data frame
}
\source{
Hamilton (2020) Optimal  contributions in highly fecund species with overlapping generations
}
\usage{
fams_example
}
\description{
A data frame defining sex-age groups
}
\references{
Hamilton (2020) Optimal  contributions in highly fecund species with overlapping generations
}
\author{
Matthew Hamilton <matthewhamilton75@gmail.com>
}
\keyword{datasets}
