% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.A_bar_bb.hamilton.R
\name{get.A_bar_bb.hamilton}
\alias{get.A_bar_bb.hamilton}
\title{get.A_bar_bb.hamilton}
\usage{
get.A_bar_bb.hamilton(group_contbn, fams, A_cand_fams)
}
\value{
L
}
\description{
get.A_bar_bb.hamilton
}
