% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.A_bar_bb.meuwissen.R
\name{get.A_bar_bb.meuwissen}
\alias{get.A_bar_bb.meuwissen}
\title{get.A_bar_bb.meuwissen}
\usage{
get.A_bar_bb.meuwissen(groups, ped, A_cand_fams, J)
}
\value{
A_bar_bb

A_cand_in_A_bar_bb
}
\description{
get.A_bar_bb.meuwissen
}
