% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.c.R
\name{get.c}
\alias{get.c}
\title{Implements optimal contributions method detailed in Hamilton MG (2020) Optimal contributions in highly fecund species with overlapping generations.}
\usage{
get.c(
  C_max,
  group_contbn,
  ped,
  fams = NULL,
  candidates,
  parents_max = NA,
  method
)
}
\arguments{
\item{C_max}{constraint on the average of relationships between members of the population (numeric between 0 and 2).}

\item{group_contbn}{a data frame defining sex-age groups with the following columns:
\itemize{
 \item{\strong{sex}} {- 'male' or 'female'.}
 \item{\strong{age}} {age group (integer greater than 0).}
 \item{\strong{fams_n}} {number of families in sex-age group (integer greater than 0).}
}}

\item{ped}{a data frame with the following columns:
\itemize{
 \item{\strong{id}} {- individual identifier (character).}
 \item{\strong{sire}} {- sire identifier (character). '0' or NA if unknown.}
 \item{\strong{dam}} {- dam identifier (character). '0' or NA if unknown.}
 \item{\strong{ebv}} {- estimated breeding value of individual (numeric)}
 \item{\strong{selection_round}} {- selection round in which individual was generated (sequential integer).  Not required if method = 'Hamilton'.}
 \item{\strong{sex}} {- 'male' or 'female'. Not required if method = 'Hamilton' and may be excluded if method = 'Meuwissen'.}
}}

\item{fams}{Must be NULL if method = 'Meuwissen'.  If method = 'Hamilton', a data frame with the following columns:
\itemize{
 \item{\strong{family_id}} {- full-sib family identifier (character).}
 \item{\strong{sire}} {- sire identifier (character). '0' or NA if unknown.}
 \item{\strong{dam}} {- dam identifier (character). '0' or NA if unknown.}
 \item{\strong{selection_round}} {- selection round in which individual was generated (sequential integer).  Not required if method = 'Hamilton'.}
}}

\item{candidates}{\itemize{
\item{\strong{id}} {- individual identifier (character).}
\item{\strong{cprev_fam}} {number of families that the individual has previously contributed to (integer greater than or equal to 0).}
\item{\strong{cmax_fam}} {maximum number of families that the individual can contribute to (integer greater than or equal to 0).}
\item{\strong{cmin_fam}} {maximum number of families that the individual can contribute to (integer greater than or equal to 0).}
\item{\strong{ebv}} {- estimated breeding value of individual (numeric)}
\item{\strong{sex}} {- 'male' or 'female'. Not required if method = 'Hamilton' and may be excluded if method = 'Meuwissen'.}
\item{\strong{age}} {age of individual (integer greater than 0).}
}}

\item{parents_max}{maximum number of parents of each sex that can be selected as a parent from any given family.  cprev_fam and cmin_fam in the candidates data frame override this constraint (integer greater than 2 or NA).}

\item{method}{optimal contribution method to implement ('Hamilton' or 'Meuwissen') - refer to Hamilton (2020) and Meuwissen and Sonesson (1998)}
}
\value{
\strong{c_1} Optimal contributions of Class 1 individuals (vector)

\strong{c_2} Contributions of Class 2 individuals (vector)

\strong{contbn_realised} Contributions of all individuals by sex-age group

\strong{A_bar_bb} Average relationships among existing sex-age groups

\strong{C_t_plus_1} Average relationship at time t+1

\strong{EBV_avg} Weighted mean estimated breeding value given c_1 and c_2

\strong{Lbar_male} Realised average generation interval for males

\strong{Lbar_female} Realised average generation interval for females

\strong{group_contbn} Data frame containing inputted group_contbn, sex_age group, s vector and r vector

\strong{ped} Pedigree file (with EBVs) for candidates and their ancestors

\strong{J} a matrix with rows representing all individuals in age groups 2 to q at time t+1 and columns representing the corresponding sex-age groups

\strong{A_1b} between-individual additive genetic relationship matrix of dimensions equal to the number of Class 1 candidates by the number of existing individuals to be retained

\strong{A_2b} between-individual additive genetic relationship matrix of dimensions equal to the number of Class 2 candidates by the number of existing individuals to be retained

\strong{M_1} M matrix of dimensions equal to the number of Class 1 candidates by columns representing sex-age groups

\strong{M_2} M matrix of dimensions equal to the number of Class 2 candidates by columns representing sex-age groups
}
\description{
get.c
}
\details{
Implements optimal contributions method detailed in Hamilton MG (2020) Optimal  contributions in highly fecund species with overlapping generations.
}
\examples{
#Retrieve data for example from Hamilton 2020 (supplementary materials)
data(group_contribution_example)
data(ped_example)
data(fams_example)
data(candidates_example)

#Run example from Hamilton 2020 (supplementary materials), implementing method = "Hamilton".
get.c(C_max = 0.45,
      group_contbn = group_contribution_example,
      ped = ped_example[,c('id', 'sire', 'dam', 'ebv')],
      fams = fams_example,
      candidates = candidates_example,
      parents_max = NA,
      method = "Hamilton")

#Run example from Hamilton 2020 (supplementary materials), implementing method = "Meuwissen".
get.c(C_max = 0.45,
      group_contbn = group_contribution_example,
      ped = ped_example,
      fams = NULL,
      candidates = candidates_example,
      parents_max = NA,
      method = "Meuwissen")
}
\references{
Hamilton MG (2020) Optimal  contributions in highly fecund species with overlapping generations

Meuwissen THE (1997) Maximizing the response of selection with a predefined rate of inbreeding. J Anim Sci 75:934-940

Meuwissen THE, Sonesson AK (1998) Maximizing the response of selection with a predefined rate of inbreeding: Overlapping generations. J Anim Sci 76:2575-2583
}
