% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.D.and.Z.R
\name{get.D.and.Z}
\alias{get.D.and.Z}
\title{Generates D and Z matrices}
\usage{
get.D.and.Z(ped, A_parents)
}
\arguments{
\item{ped}{A three column pedigree file including candidate individuals only}

\item{A_parents}{A numerator relationship matrix including parents only}
}
\value{
Z

D_diag

D_inv_diag

F_bar
}
\description{
Generates D and Z matrices
}
