% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.M.meuwissen.R
\name{get.M.meuwissen}
\alias{get.M.meuwissen}
\title{get.M.meuwissen}
\usage{
get.M.meuwissen(ped_b, A_cand_fams, c1_candidates, c2_candidates, group_contbn)
}
\value{
J

A_1b

A_2b

M_1

M_2
}
\description{
get.M.meuwissen
}
