% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_contribution_example.R
\docType{data}
\name{group_contribution_example}
\alias{group_contribution_example}
\title{ped_example data for small worked example in Hamilton (2020) Optimal  contributions in highly fecund species with overlapping generations}
\format{
Data frame
}
\source{
Hamilton (2020) Optimal  contributions in highly fecund species with overlapping generations
}
\usage{
group_contribution_example
}
\description{
A data frame defining sex-age groups
}
\references{
Hamilton (2020) Optimal  contributions in highly fecund species with overlapping generations
}
\author{
Matthew Hamilton <matthewhamilton75@gmail.com>
}
\keyword{datasets}
