% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.matrix.R
\name{incidence.matrix}
\alias{incidence.matrix}
\title{Generates an incidence matrix}
\usage{
incidence.matrix(row_values, row_names, col_names)
}
\arguments{
\item{row_values}{vector of values equal in length to row_names.  Element values must be present in col_names.}

\item{row_names}{vector of row names}

\item{col_names}{vector of column names}
}
\value{
An incidence matrix with 1 in elements where row_values match col_names
}
\description{
Generates an incidence matrix
}
