% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Checks if a numeric input is a whole number}
\usage{
is.wholenumber(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
TRUE if x is a whole number and FALSE if it is not.
}
\description{
Checks if a numeric input is a whole number
}
