% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.ped.R
\name{reduce.ped}
\alias{reduce.ped}
\title{Removes unncessary rows from a 3-column pedigree file by retaining only ancestors of specified individuals}
\usage{
reduce.ped(ped, indivs)
}
\arguments{
\item{ped}{A three column pedigree file including candidate individuals only}

\item{indivs}{List of individuals for which ancestors will be retained}
}
\value{
A a 3-column pedigree file retaining only ancestors of specified individuals (and the individuals themselves).
}
\description{
Removes unncessary rows from a 3-column pedigree file by retaining only ancestors of specified individuals
}
